+++ Nos/AppController.swift
@@ -24,8 +24,6 @@ class AppController: ObservableObject {
     @Dependency(\.currentUser) private var currentUser
     
     init() {
-        print("Initializing analytics. This is hack to get the dependency initialzed by printing it: \(analytics)")
-        self.currentUser = currentUser
         Log.info("App Version: \(Bundle.current.versionAndBuild)")
     }
