+++ Nos/Extensions/Foundation+Sendable.swift
@@ -1,0 +1,11 @@
+//
+//  Foundation+Sendable.swift
+//  Nos
+//
+//  Created by Matthew Lorentz on 4/18/23.
+//
+
+import Foundation
+
+extension UserDefaults: @unchecked Sendable {}
+extension AttributedString: @unchecked Sendable {}
