+++ Nos/NosApp.swift
@@ -9,45 +9,6 @@ import SwiftUI
 import Logger
 import Dependencies
 
-private enum CurrentUserKey: DependencyKey {
-    static let liveValue = CurrentUser.shared
-    static let testValue = CurrentUser.shared
-    static let previewValue = CurrentUser.shared
-}
-
-extension DependencyValues {
-    var currentUser: CurrentUser {
-        get { self[CurrentUserKey.self] }
-        set { self[CurrentUserKey.self] = newValue }
-    }
-}
-
-private enum RouterKey: DependencyKey {
-    static let liveValue = Router()
-    static let testValue = Router()
-    static let previewValue = Router()
-}
-
-extension DependencyValues {
-    var router: Router {
-        get { self[RouterKey.self] }
-        set { self[RouterKey.self] = newValue }
-    }
-}
-
-private enum RelayServiceKey: DependencyKey {
-    static let liveValue = RelayService(persistenceController: PersistenceController.shared)
-    static let testValue = RelayService(persistenceController: PersistenceController.shared)
-    static let previewValue = RelayService(persistenceController: PersistenceController.shared)
-}
-
-extension DependencyValues {
-    var relayService: RelayService {
-        get { self[RelayServiceKey.self] }
-        set { self[RelayServiceKey.self] = newValue }
-    }
-}
-
 @main
 struct NosApp: App {
